test(sysrap): add SGLFW_test
mv sysrap/SGLFW_tests/SGLFW_tests.cc sysrap/tests/
